d/p/systemd-socket-activation.patch: do not leak sockets in child process
This fixes two things:

First is that the close_listen_socks() call in the child process
resulting from the fork() in server_accept_loop() does need to force
close the listen socks. Otherwise, the child process maintains an
unnecessary listener that will not be closed when ssh.service stops,
hence causing errors when ssh.service starts again.

Second is that we force rexec_flag = 0 when sshd is socket-activated so
that the child process does not re-exec. In effect with the current
logic, this makes sure that we do not get a duplicate socket for the
active ssh session.